# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Merchants API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/merchants/fake-login": {
            "post": {
                "tags": [
                    "Merchants"
                ],
                "summary": "Login Comercio",
                "operationId": "25a3382d48bef9df11234765d4f4964a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comercio creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos no validos"
                    }
                }
            }
        },
        "/api/merchants": {
            "get": {
                "tags": [
                    "Merchants"
                ],
                "summary": "Obtiene todos los Comercios",
                "operationId": "29dc2feea9d224f1903cfe3d27bd81fb",
                "responses": {
                    "200": {
                        "description": "Lista de Comercios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Merchant"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Merchants"
                ],
                "summary": "Crear un nuevo Comercio",
                "operationId": "15db992a94c43f36e50daf7853b53ca2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Merchant"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comercio creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Merchant"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos no validos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/merchants/{id}": {
            "get": {
                "tags": [
                    "Merchants"
                ],
                "summary": "Obtener un Comercio por ID",
                "operationId": "56beef28bbe1d805bab17fb72672742e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del Comercio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comercio encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Merchant"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comercio no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Merchants"
                ],
                "summary": "Actualizar un Comercio",
                "operationId": "1bfd0f2f913af1bcfbe7306aafd7e5ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del Comercio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Merchant"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comercio actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Merchant"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos no validos"
                    },
                    "404": {
                        "description": "Comercio no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Merchants"
                ],
                "summary": "Eliminar un Comercio",
                "operationId": "b4a0d51bc1209b6cb6c808e902595592",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del Comercio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Comercio eliminado exitosamente"
                    },
                    "404": {
                        "description": "Comercio no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payments": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Obtiene todos los Pagos",
                "operationId": "a2236c2235ed8954c443cd2dd269bea1",
                "responses": {
                    "200": {
                        "description": "Lista de Pagos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Payment"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Crear un nuevo Pago",
                "operationId": "5e907b772c6a27dc7dafc6c76a5ffa16",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Payment"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pago creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos no validos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payments/{id}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Obtener un Pago por ID",
                "operationId": "30f448f98368b8382bd347740dbb9fab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del Pago",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pago encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pago no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Payments"
                ],
                "summary": "Actualizar un Pago",
                "operationId": "4520afc2aa94b91f45e0d86129fee420",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del Pago",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Payment"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pago actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Payments"
                ],
                "summary": "Eliminar un Pago",
                "operationId": "41170d6425c9fe8a73fbe9b718627e59",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del Pago",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Pago eliminado exitosamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Merchant": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "johndoe@example.com"
                    }
                },
                "type": "object"
            },
            "Payment": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "merchant_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "amount": {
                        "type": "number",
                        "format": "decimal",
                        "example": 19
                    },
                    "method": {
                        "type": "string",
                        "enum": [
                            "tarjeta_credito",
                            "pix",
                            "transferencia"
                        ],
                        "example": "tarjeta_credito"
                    },
                    "fee": {
                        "type": "number",
                        "format": "decimal",
                        "example": 2
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pendiente",
                            "procesado",
                            "fallido"
                        ],
                        "example": "pendiente"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Merchants",
            "description": "Merchants"
        },
        {
            "name": "Payments",
            "description": "Payments"
        }
    ],
    "security": {
        "1": {
            "bearerAuth": []
        }
    }
}